+++ src/world.rs
@@ -32,13 +32,13 @@ impl World {
 
 impl Hitable for World {
     /// Determine if the given ray has hit any of the objects in the world
-    fn hit(&self, ray: &Ray, position_min: f64, position_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, position_min: f32, position_max: f32) -> Option<HitRecord> {
         let mut record = HitRecord::new(0.0, 0.0, 0.0,
                                         Vector3::zeros(),
                                         Vector3::zeros(),
                                         Box::new(Diffuse::new(ConstantTexture::new(0.0, 0.0, 0.0))));
         let mut hit_anything: bool = false;
-        let mut closed_so_far: f64 = position_max;
+        let mut closed_so_far: f32 = position_max;
 
         for object in &self.objects {
             match object.hit(ray, position_min, closed_so_far) {
